test: add tests for analytics-connector
